# Conflicts:
#	target/classes/org/control_parental/comentario/domain/ComentarioService.class
#	target/test-classes/org/control_parental/hijo/application/HijoControllerIntegrationTest.class
#	target/test-classes/org/control_parental/padre/application/PadreControllerTest.class
#	target/test-classes/org/control_parental/salon/application/SalonControllerIntegrationTest.class

+++ src/main/java/org/control_parental/publicacion/domain/PublicacionService.java
@@ -11,9 +11,9 @@ import org.control_parental.hijo.infrastructure.HijoRepository;
 import org.control_parental.like.Domain.Padre_Like;
 import org.control_parental.like.Infrastructure.LikeRepository;
 import org.control_parental.padre.domain.Padre;
+import org.control_parental.padre.dto.PadrePublicacionDto;
 import org.control_parental.padre.infrastructure.PadreRepository;
 import org.control_parental.profesor.domain.Profesor;
-import org.control_parental.profesor.dto.ProfesorResponseDto;
 import org.control_parental.profesor.infrastructure.ProfesorRepository;
 import org.control_parental.publicacion.dto.NewPublicacionDto;
 import org.control_parental.publicacion.dto.PublicacionResponseDto;
@@ -25,6 +25,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 
+import java.nio.file.AccessDeniedException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -70,7 +71,7 @@ public class PublicacionService {
 //        newPublicacion.setFecha(LocalDateTime.now());
 //        newPublicacion.setLikes(0);
 
-    public void savePublicacion(NewPublicacionDto newPublicacionDto) {
+    public String savePublicacion(NewPublicacionDto newPublicacionDto) {
         //obtener quien lo esta publicando con Sprnig Scurity
         String email = authorizationUtils.authenticateUser();
 
@@ -80,7 +81,7 @@ public class PublicacionService {
         Salon salon = salonRepository.findById(newPublicacionDto.getSalonId()).orElseThrow(
                 ()-> new ResourceNotFoundException("El salon no existe"));
 
-        List<Hijo> hijos = new ArrayList<Hijo>();
+        List<Hijo> hijos = new ArrayList<>();
         newPublicacion.setTitulo(newPublicacionDto.getTitulo());
         newPublicacion.setDescripcion(newPublicacionDto.getDescripcion());
         newPublicacion.setFecha(LocalDateTime.now());
@@ -108,13 +109,13 @@ public class PublicacionService {
         profesor.getPublicaciones().add(newPublicacion);
         salon.getPublicaciones().add(newPublicacion);
         publicacionRepository.save(newPublicacion);
+        return "/" + newPublicacion.getId();
     }
 
     public PublicacionResponseDto getPublicacionById(Long id) {
         Publicacion publicacion = publicacionRepository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException("Esta publicacion no existe"));
         PublicacionResponseDto publicacionResponseDto = modelMapper.map(publicacion, PublicacionResponseDto.class);
-        publicacionResponseDto.setProfesor(modelMapper.map(publicacion.getProfesor(), ProfesorResponseDto.class));
 
         List<HijoPublicacionDto> hijoPublicacionDtos = new ArrayList<>();
         publicacion.getHijos().forEach((hijo) -> {
@@ -126,18 +127,26 @@ public class PublicacionService {
             comentarioPublicacionDtos.add(modelMapper.map(comentario, ComentarioPublicacionDto.class));
         });
 
+        List<PadrePublicacionDto> padrePublicacionDtos = new ArrayList<>();
+        publicacion.getLikers().forEach((liker) -> {
+            padrePublicacionDtos.add(modelMapper.map(liker.getPadre(), PadrePublicacionDto.class));
+        });
+
         publicacionResponseDto.setHijos(hijoPublicacionDtos);
         publicacionResponseDto.setComentarios(comentarioPublicacionDtos);
+        publicacionResponseDto.setLikers(padrePublicacionDtos);
         return publicacionResponseDto;
     }
 
-    public void deletePublicacion(Long id)
-    {
+    public void deletePublicacion(Long id) throws AccessDeniedException {
         String email = authorizationUtils.authenticateUser();
+        Long userId = publicacionRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Publicación no encontrada")).getProfesor().getId();
+        authorizationUtils.verifyUserAuthorization(email, userId);
+
         publicacionRepository.deleteById(id);
     }
-/*
-    public void patchPublicacion(Long id, NewPublicacionDto newPublicacionDto){
+
+    /*public void patchPublicacion(Long id, NewPublicacionDto newPublicacionDto){
         Publicacion publicacion = publicacionRepository.findById(id).orElseThrow();
         publicacion.setFoto(newPublicacionDto.getFoto());
         publicacion.setTitulo(newPublicacionDto.getTitulo());
@@ -145,8 +154,8 @@ public class PublicacionService {
         List<Hijo> hijos = new ArrayList<Hijo>();
         newPublicacionDto.getHijos_id().forEach(hijoId -> {hijos.add(hijoRepository.findById(hijoId).orElseThrow());});
         publicacion.setHijos(hijos);
-    }
-*/
+    }*/
+
     public List<PublicacionResponseDto> findPostsBySalonId(Long salon_id) {
         Salon salon = salonRepository.findById(salon_id).orElseThrow();
         List<Publicacion> publicaciones = publicacionRepository.findAllBySalon(salon);
@@ -190,6 +199,10 @@ public class PublicacionService {
         Padre_Like like = likeRepository.findByPadre_IdAndPublicacion_Id(padre.getId(), postId).orElseThrow(
                 ()-> new ResourceNotFoundException("Esta persona no ha dado like a esta publicacion")
         );
+        Publicacion publicacion = publicacionRepository.findById(postId).orElseThrow(() -> new ResourceNotFoundException("Publicación no encontrada"));
+        publicacion.setLikes(publicacion.getLikes() - 1);
+        publicacionRepository.save(publicacion);
         likeRepository.delete(like);
     }
+
 }
